refactor: move AtomNavigationThrottle out of net folder
